Fix IO bug with extra field in the metadata and add test.
